requirements.txt
alembic==1.8.1
anyio==3.6.1
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.0.1
fastapi==0.85.1
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
Mako==1.2.3
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
# psycopg2-binary==2.9.5 ## uncomment if ruuning on Apple M1 architecture. For deploying on Linux machines for example, psycopg2 will be installed as well as libpq from Dockerfile.
pyasn1==0.4.8
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
